refactor: split prompt template detection
