+++ crates/bitnet-prompt-templates-core/src/lib.rs
@@ -6,6 +6,8 @@
 #![allow(rustdoc::broken_intra_doc_links)]
 #![allow(rustdoc::invalid_html_tags)]
 
+mod detection;
+
 use anyhow::{Result, bail};
 use serde::{Deserialize, Serialize};
 use std::path::Path;
@@ -506,770 +508,7 @@ impl TemplateType {
     /// 2. Tokenizer family name heuristics
     /// 3. Fallback to Raw
     pub fn detect(tokenizer_name: Option<&str>, chat_template_jinja: Option<&str>) -> Self {
-        // Priority 1: GGUF chat_template metadata
-        if let Some(jinja) = chat_template_jinja {
-            // LLaMA-3 signature
-            if jinja.contains("<|start_header_id|>") && jinja.contains("<|eot_id|>") {
-                tracing::debug!(
-                    template = "Llama3Chat",
-                    source = "gguf_chat_template",
-                    "auto-detected prompt template"
-                );
-                return Self::Llama3Chat;
-            }
-            if Self::looks_like_bitnet_answer_template(jinja) {
-                tracing::debug!(
-                    template = "BitnetCppAnswer",
-                    source = "gguf_chat_template",
-                    "auto-detected prompt template"
-                );
-                return Self::BitnetCppAnswer;
-            }
-            // Fill-in-the-middle signature
-            if jinja.contains("<fim_prefix>") {
-                tracing::debug!(
-                    template = "FillInMiddle",
-                    source = "gguf_chat_template",
-                    "auto-detected prompt template"
-                );
-                return Self::FillInMiddle;
-            }
-            // Command-R+ signature (must be before CohereCommand)
-            if jinja.contains("<|START_OF_TURN_TOKEN|>") {
-                tracing::debug!(
-                    template = "CommandRPlus",
-                    source = "gguf_chat_template",
-                    "auto-detected prompt template"
-                );
-                return Self::CommandRPlus;
-            }
-            // Granite signature (must be before ChatML check)
-            if jinja.contains("<|start_of_role|>") {
-                tracing::debug!(
-                    template = "GraniteChat",
-                    source = "gguf_chat_template",
-                    "auto-detected prompt template"
-                );
-                return Self::GraniteChat;
-            }
-            // Nemotron signature (must be before ChatML check)
-            if jinja.contains("<extra_id_0>") || jinja.contains("<extra_id_1>") {
-                tracing::debug!(
-                    template = "NemotronChat",
-                    source = "gguf_chat_template",
-                    "auto-detected prompt template"
-                );
-                return Self::NemotronChat;
-            }
-            // Phi-3 signature (must be before ChatML check; uses <|system|>/<|end|>/<|user|>)
-            if jinja.contains("<|system|>")
-                && jinja.contains("<|end|>")
-                && jinja.contains("<|user|>")
-            {
-                tracing::debug!(
-                    template = "Phi3Instruct",
-                    source = "gguf_chat_template",
-                    "auto-detected prompt template"
-                );
-                return Self::Phi3Instruct;
-            }
-            // EXAONE signature (must be before ChatML check)
-            if jinja.contains("[|system|]") || jinja.contains("[|endofturn|]") {
-                tracing::debug!(
-                    template = "ExaoneChat",
-                    source = "gguf_chat_template",
-                    "auto-detected prompt template"
-                );
-                return Self::ExaoneChat;
-            }
-            // ChatML / Phi-4 signature
-            if jinja.contains("<|im_start|>") && jinja.contains("<|im_end|>") {
-                tracing::debug!(
-                    template = "Phi4Chat",
-                    source = "gguf_chat_template",
-                    "auto-detected prompt template"
-                );
-                return Self::Phi4Chat;
-            }
-            // Gemma signature
-            if jinja.contains("<start_of_turn>") && jinja.contains("<end_of_turn>") {
-                tracing::debug!(
-                    template = "GemmaChat",
-                    source = "gguf_chat_template",
-                    "auto-detected prompt template"
-                );
-                return Self::GemmaChat;
-            }
-            // [INST] with <<SYS>>/<</SYS>> ΓåÆ Llama2Chat (CodeLlama handled by tokenizer name)
-            if jinja.contains("[INST]") && jinja.contains("<<SYS>>") && jinja.contains("<</SYS>>") {
-                tracing::debug!(
-                    template = "Llama2Chat",
-                    source = "gguf_chat_template",
-                    "auto-detected prompt template"
-                );
-                return Self::Llama2Chat;
-            }
-            // Mistral [INST] signature (no <<SYS>>)
-            if jinja.contains("[INST]") && jinja.contains("[/INST]") {
-                tracing::debug!(
-                    template = "MistralChat",
-                    source = "gguf_chat_template",
-                    "auto-detected prompt template"
-                );
-                return Self::MistralChat;
-            }
-            // Cohere Command signature
-            if jinja.contains("<|START_OF_TURN_TOKEN|>") && jinja.contains("<|END_OF_TURN_TOKEN|>")
-            {
-                tracing::debug!(
-                    template = "CohereCommand",
-                    source = "gguf_chat_template",
-                    "auto-detected prompt template"
-                );
-                return Self::CohereCommand;
-            }
-            // ChatGLM/GLM-4 signature (requires [gMASK])
-            if jinja.contains("[gMASK]") {
-                tracing::debug!(
-                    template = "ChatGLMChat",
-                    source = "gguf_chat_template",
-                    "auto-detected prompt template"
-                );
-                return Self::ChatGLMChat;
-            }
-            // Zephyr signature (</s> delimiters with <|user|>, no [gMASK] or <|im_start|>)
-            if jinja.contains("</s>")
-                && jinja.contains("<|user|>")
-                && !jinja.contains("[gMASK]")
-                && !jinja.contains("<|im_start|>")
-            {
-                tracing::debug!(
-                    template = "ZephyrChat",
-                    source = "gguf_chat_template",
-                    "auto-detected prompt template"
-                );
-                return Self::ZephyrChat;
-            }
-            // OLMo instruct signature (<|user|>/<|assistant|> without [gMASK])
-            if jinja.contains("<|user|>") && jinja.contains("<|assistant|>") {
-                tracing::debug!(
-                    template = "OlmoInstruct",
-                    source = "gguf_chat_template",
-                    "auto-detected prompt template"
-                );
-                return Self::OlmoInstruct;
-            }
-            // Alpaca ### Instruction/### Response markers (but NOT ### User: which is Solar)
-            if jinja.contains("### Instruction:")
-                && jinja.contains("### Response:")
-                && !jinja.contains("### User:")
-            {
-                tracing::debug!(
-                    template = "AlpacaInstruct",
-                    source = "gguf_chat_template",
-                    "auto-detected prompt template"
-                );
-                return Self::AlpacaInstruct;
-            }
-            // Solar ### User:/### Assistant: markers
-            if jinja.contains("### User:") && jinja.contains("### Assistant:") {
-                tracing::debug!(
-                    template = "SolarInstruct",
-                    source = "gguf_chat_template",
-                    "auto-detected prompt template"
-                );
-                return Self::SolarInstruct;
-            }
-            // MPT ### instruction markers
-            if jinja.contains("### Instruction") && jinja.contains("### Response") {
-                tracing::debug!(
-                    template = "MptInstruct",
-                    source = "gguf_chat_template",
-                    "auto-detected prompt template"
-                );
-                return Self::MptInstruct;
-            }
-            // OpenChat GPT4 Correct signature
-            if jinja.contains("GPT4 Correct") {
-                tracing::debug!(
-                    template = "OpenChat",
-                    source = "gguf_chat_template",
-                    "auto-detected prompt template"
-                );
-                return Self::OpenChat;
-            }
-            // WizardLM signature (USER:/ASSISTANT: with "A chat between")
-            if jinja.contains("USER:")
-                && jinja.contains("ASSISTANT:")
-                && jinja.contains("A chat between")
-            {
-                tracing::debug!(
-                    template = "WizardLM",
-                    source = "gguf_chat_template",
-                    "auto-detected prompt template"
-                );
-                return Self::WizardLM;
-            }
-            // Vicuna/ShareGPT signature (USER:/ASSISTANT:, not Falcon's "User:")
-            if jinja.contains("USER:") && jinja.contains("ASSISTANT:") {
-                tracing::debug!(
-                    template = "VicunaChat",
-                    source = "gguf_chat_template",
-                    "auto-detected prompt template"
-                );
-                return Self::VicunaChat;
-            }
-            // Generic instruct template
-            if jinja.contains("{% for message in messages %}") {
-                tracing::debug!(
-                    template = "Instruct",
-                    source = "gguf_chat_template",
-                    "auto-detected prompt template"
-                );
-                return Self::Instruct;
-            }
-        }
-
-        // Priority 2: Tokenizer family name heuristics
-        if let Some(name) = tokenizer_name {
-            let lower = name.to_ascii_lowercase();
-            // TinyLlama must be checked before "llama" to avoid false match
-            if lower.contains("tinyllama") || lower.contains("tiny-llama") {
-                tracing::debug!(
-                    template = "TinyLlamaChat",
-                    source = "tokenizer_name",
-                    hint = name,
-                    "auto-detected prompt template"
-                );
-                return Self::TinyLlamaChat;
-            }
-            if lower.contains("llama-3.2")
-                || lower.contains("llama3.2")
-                || lower.contains("llama-32")
-            {
-                tracing::debug!(
-                    template = "Llama32Chat",
-                    source = "tokenizer_name",
-                    hint = name,
-                    "auto-detected prompt template"
-                );
-                return Self::Llama32Chat;
-            }
-            if lower.contains("llama-3.1")
-                || lower.contains("llama3.1")
-                || lower.contains("llama-31")
-            {
-                tracing::debug!(
-                    template = "Llama31Chat",
-                    source = "tokenizer_name",
-                    hint = name,
-                    "auto-detected prompt template"
-                );
-                return Self::Llama31Chat;
-            }
-            if lower.contains("llama3") || lower.contains("llama-3") {
-                tracing::debug!(
-                    template = "Llama3Chat",
-                    source = "tokenizer_name",
-                    hint = name,
-                    "auto-detected prompt template"
-                );
-                return Self::Llama3Chat;
-            }
-            if lower.contains("llama2") || lower.contains("llama-2") {
-                tracing::debug!(
-                    template = "Llama2Chat",
-                    source = "tokenizer_name",
-                    hint = name,
-                    "auto-detected prompt template"
-                );
-                return Self::Llama2Chat;
-            }
-            if lower.contains("qwen2.5") || lower.contains("qwen-2.5") {
-                tracing::debug!(
-                    template = "Qwen25Chat",
-                    source = "tokenizer_name",
-                    hint = name,
-                    "auto-detected prompt template"
-                );
-                return Self::Qwen25Chat;
-            }
-            if lower.contains("qwen") {
-                tracing::debug!(
-                    template = "QwenChat",
-                    source = "tokenizer_name",
-                    hint = name,
-                    "auto-detected prompt template"
-                );
-                return Self::QwenChat;
-            }
-            if lower.contains("phi2") || lower.contains("phi-2") {
-                tracing::debug!(
-                    template = "Phi2Instruct",
-                    source = "tokenizer_name",
-                    hint = name,
-                    "auto-detected prompt template"
-                );
-                return Self::Phi2Instruct;
-            }
-            if lower.contains("phi3") || lower.contains("phi-3") {
-                tracing::debug!(
-                    template = "Phi3Instruct",
-                    source = "tokenizer_name",
-                    hint = name,
-                    "auto-detected prompt template"
-                );
-                return Self::Phi3Instruct;
-            }
-            // Dolphin must be checked before "phi" because "dolphin" contains "phi"
-            if lower.contains("dolphin") {
-                tracing::debug!(
-                    template = "DolphinChat",
-                    source = "tokenizer_name",
-                    hint = name,
-                    "auto-detected prompt template"
-                );
-                return Self::DolphinChat;
-            }
-            if lower.contains("phi") {
-                tracing::debug!(
-                    template = "Phi4Chat",
-                    source = "tokenizer_name",
-                    hint = name,
-                    "auto-detected prompt template"
-                );
-                return Self::Phi4Chat;
-            }
-            if lower.contains("codegemma") || lower.contains("code-gemma") {
-                tracing::debug!(
-                    template = "CodeGemma",
-                    source = "tokenizer_name",
-                    hint = name,
-                    "auto-detected prompt template"
-                );
-                return Self::CodeGemma;
-            }
-            if lower.contains("gemma2") || lower.contains("gemma-2-") {
-                tracing::debug!(
-                    template = "Gemma2Chat",
-                    source = "tokenizer_name",
-                    hint = name,
-                    "auto-detected prompt template"
-                );
-                return Self::Gemma2Chat;
-            }
-            if lower.contains("gemma") {
-                tracing::debug!(
-                    template = "GemmaChat",
-                    source = "tokenizer_name",
-                    hint = name,
-                    "auto-detected prompt template"
-                );
-                return Self::GemmaChat;
-            }
-            if lower.contains("mixtral") {
-                tracing::debug!(
-                    template = "MixtralInstruct",
-                    source = "tokenizer_name",
-                    hint = name,
-                    "auto-detected prompt template"
-                );
-                return Self::MixtralInstruct;
-            }
-            if lower.contains("mistral-nemo") || lower.contains("nemo") {
-                tracing::debug!(
-                    template = "MistralNemoChat",
-                    source = "tokenizer_name",
-                    hint = name,
-                    "auto-detected prompt template"
-                );
-                return Self::MistralNemoChat;
-            }
-            if lower.contains("mistral") {
-                tracing::debug!(
-                    template = "MistralChat",
-                    source = "tokenizer_name",
-                    hint = name,
-                    "auto-detected prompt template"
-                );
-                return Self::MistralChat;
-            }
-            if lower.contains("deepseek-v3")
-                || lower.contains("deepseekv3")
-                || lower.contains("deepseek3")
-            {
-                tracing::debug!(
-                    template = "DeepSeekV3Chat",
-                    source = "tokenizer_name",
-                    hint = name,
-                    "auto-detected prompt template"
-                );
-                return Self::DeepSeekV3Chat;
-            }
-            if lower.contains("deepseek") {
-                tracing::debug!(
-                    template = "DeepSeekChat",
-                    source = "tokenizer_name",
-                    hint = name,
-                    "auto-detected prompt template"
-                );
-                return Self::DeepSeekChat;
-            }
-            if lower.contains("fim") || lower.contains("fill-in-middle") {
-                tracing::debug!(
-                    template = "FillInMiddle",
-                    source = "tokenizer_name",
-                    hint = name,
-                    "auto-detected prompt template"
-                );
-                return Self::FillInMiddle;
-            }
-            if lower.contains("starcoder") || lower.contains("star-coder") {
-                tracing::debug!(
-                    template = "StarCoder",
-                    source = "tokenizer_name",
-                    hint = name,
-                    "auto-detected prompt template"
-                );
-                return Self::StarCoder;
-            }
-            if lower.contains("falcon-2") || lower.contains("falcon2") {
-                tracing::debug!(
-                    template = "Falcon2Chat",
-                    source = "tokenizer_name",
-                    hint = name,
-                    "auto-detected prompt template"
-                );
-                return Self::Falcon2Chat;
-            }
-            if lower.contains("falcon") {
-                tracing::debug!(
-                    template = "FalconChat",
-                    source = "tokenizer_name",
-                    hint = name,
-                    "auto-detected prompt template"
-                );
-                return Self::FalconChat;
-            }
-            if lower.contains("codellama") || lower.contains("code-llama") {
-                tracing::debug!(
-                    template = "CodeLlamaInstruct",
-                    source = "tokenizer_name",
-                    hint = name,
-                    "auto-detected prompt template"
-                );
-                return Self::CodeLlamaInstruct;
-            }
-            if lower.contains("aya") {
-                tracing::debug!(
-                    template = "CohereAya",
-                    source = "tokenizer_name",
-                    hint = name,
-                    "auto-detected prompt template"
-                );
-                return Self::CohereAya;
-            }
-            if lower.contains("cohere") || lower.contains("command-r") {
-                tracing::debug!(
-                    template = "CommandRPlus",
-                    source = "tokenizer_name",
-                    hint = name,
-                    "auto-detected prompt template"
-                );
-                return Self::CommandRPlus;
-            }
-            if lower.contains("internlm") {
-                tracing::debug!(
-                    template = "InternLMChat",
-                    source = "tokenizer_name",
-                    hint = name,
-                    "auto-detected prompt template"
-                );
-                return Self::InternLMChat;
-            }
-            if lower.contains("yi-") || lower.contains("yi_") || lower == "yi" {
-                tracing::debug!(
-                    template = "YiChat",
-                    source = "tokenizer_name",
-                    hint = name,
-                    "auto-detected prompt template"
-                );
-                return Self::YiChat;
-            }
-            if lower.contains("baichuan") {
-                tracing::debug!(
-                    template = "BaichuanChat",
-                    source = "tokenizer_name",
-                    hint = name,
-                    "auto-detected prompt template"
-                );
-                return Self::BaichuanChat;
-            }
-            if lower.contains("chatglm") || lower.contains("glm-4") || lower.contains("glm4") {
-                tracing::debug!(
-                    template = "ChatGLMChat",
-                    source = "tokenizer_name",
-                    hint = name,
-                    "auto-detected prompt template"
-                );
-                return Self::ChatGLMChat;
-            }
-            if lower.contains("mpt") {
-                tracing::debug!(
-                    template = "MptInstruct",
-                    source = "tokenizer_name",
-                    hint = name,
-                    "auto-detected prompt template"
-                );
-                return Self::MptInstruct;
-            }
-            if lower.contains("rwkv") {
-                tracing::debug!(
-                    template = "RwkvWorld",
-                    source = "tokenizer_name",
-                    hint = name,
-                    "auto-detected prompt template"
-                );
-                return Self::RwkvWorld;
-            }
-            if lower.contains("olmo-2") || lower.contains("olmo2") {
-                tracing::debug!(
-                    template = "OLMo2Chat",
-                    source = "tokenizer_name",
-                    hint = name,
-                    "auto-detected prompt template"
-                );
-                return Self::OLMo2Chat;
-            }
-            if lower.contains("olmo") {
-                tracing::debug!(
-                    template = "OlmoInstruct",
-                    source = "tokenizer_name",
-                    hint = name,
-                    "auto-detected prompt template"
-                );
-                return Self::OlmoInstruct;
-            }
-            if lower.contains("zephyr") {
-                tracing::debug!(
-                    template = "ZephyrChat",
-                    source = "tokenizer_name",
-                    hint = name,
-                    "auto-detected prompt template"
-                );
-                return Self::ZephyrChat;
-            }
-            if lower.contains("vicuna") || lower.contains("sharegpt") {
-                tracing::debug!(
-                    template = "VicunaChat",
-                    source = "tokenizer_name",
-                    hint = name,
-                    "auto-detected prompt template"
-                );
-                return Self::VicunaChat;
-            }
-            if lower.contains("orca") {
-                tracing::debug!(
-                    template = "OrcaChat",
-                    source = "tokenizer_name",
-                    hint = name,
-                    "auto-detected prompt template"
-                );
-                return Self::OrcaChat;
-            }
-            if lower.contains("solar") {
-                tracing::debug!(
-                    template = "SolarInstruct",
-                    source = "tokenizer_name",
-                    hint = name,
-                    "auto-detected prompt template"
-                );
-                return Self::SolarInstruct;
-            }
-            if lower.contains("alpaca") {
-                tracing::debug!(
-                    template = "AlpacaInstruct",
-                    source = "tokenizer_name",
-                    hint = name,
-                    "auto-detected prompt template"
-                );
-                return Self::AlpacaInstruct;
-            }
-            if lower.contains("nous") || lower.contains("hermes") {
-                tracing::debug!(
-                    template = "NousHermes",
-                    source = "tokenizer_name",
-                    hint = name,
-                    "auto-detected prompt template"
-                );
-                return Self::NousHermes;
-            }
-            if lower.contains("wizard") {
-                tracing::debug!(
-                    template = "WizardLM",
-                    source = "tokenizer_name",
-                    hint = name,
-                    "auto-detected prompt template"
-                );
-                return Self::WizardLM;
-            }
-            if lower.contains("openchat") {
-                tracing::debug!(
-                    template = "OpenChat",
-                    source = "tokenizer_name",
-                    hint = name,
-                    "auto-detected prompt template"
-                );
-                return Self::OpenChat;
-            }
-            if lower.contains("granite") {
-                tracing::debug!(
-                    template = "GraniteChat",
-                    source = "tokenizer_name",
-                    hint = name,
-                    "auto-detected prompt template"
-                );
-                return Self::GraniteChat;
-            }
-            if lower.contains("nemotron") {
-                tracing::debug!(
-                    template = "NemotronChat",
-                    source = "tokenizer_name",
-                    hint = name,
-                    "auto-detected prompt template"
-                );
-                return Self::NemotronChat;
-            }
-            if lower.contains("saiga") {
-                tracing::debug!(
-                    template = "SaigaChat",
-                    source = "tokenizer_name",
-                    hint = name,
-                    "auto-detected prompt template"
-                );
-                return Self::SaigaChat;
-            }
-            // ChatGPT/GPT-4 must be checked before generic "instruct" fallback
-            // and must NOT match "gpt2" (that's the base GPT template)
-            if lower.contains("chatgpt")
-                || lower.contains("gpt-4")
-                || (lower.contains("gpt4") && !lower.contains("gpt2"))
-            {
-                tracing::debug!(
-                    template = "ChatGptChat",
-                    source = "tokenizer_name",
-                    hint = name,
-                    "auto-detected prompt template"
-                );
-                return Self::ChatGptChat;
-            }
-            if lower.contains("stablelm")
-                || lower.contains("stable-lm")
-                || lower.contains("stablecode")
-            {
-                tracing::debug!(
-                    template = "StableLMChat",
-                    source = "tokenizer_name",
-                    hint = name,
-                    "auto-detected prompt template"
-                );
-                return Self::StableLMChat;
-            }
-            if lower.contains("bloom") || lower.contains("bloomz") {
-                tracing::debug!(
-                    template = "BloomChat",
-                    source = "tokenizer_name",
-                    hint = name,
-                    "auto-detected prompt template"
-                );
-                return Self::BloomChat;
-            }
-            if lower.contains("jamba") {
-                tracing::debug!(
-                    template = "JambaChat",
-                    source = "tokenizer_name",
-                    hint = name,
-                    "auto-detected prompt template"
-                );
-                return Self::JambaChat;
-            }
-            if lower.contains("persimmon") || lower.contains("adept") {
-                tracing::debug!(
-                    template = "PersimmonChat",
-                    source = "tokenizer_name",
-                    hint = name,
-                    "auto-detected prompt template"
-                );
-                return Self::PersimmonChat;
-            }
-            if lower.contains("xverse") {
-                tracing::debug!(
-                    template = "XverseChat",
-                    source = "tokenizer_name",
-                    hint = name,
-                    "auto-detected prompt template"
-                );
-                return Self::XverseChat;
-            }
-            if lower.contains("arctic") {
-                tracing::debug!(
-                    template = "ArcticInstruct",
-                    source = "tokenizer_name",
-                    hint = name,
-                    "auto-detected prompt template"
-                );
-                return Self::ArcticInstruct;
-            }
-            if lower.contains("dbrx") {
-                tracing::debug!(
-                    template = "DbrxInstruct",
-                    source = "tokenizer_name",
-                    hint = name,
-                    "auto-detected prompt template"
-                );
-                return Self::DbrxInstruct;
-            }
-            if lower.contains("exaone") {
-                tracing::debug!(
-                    template = "ExaoneChat",
-                    source = "tokenizer_name",
-                    hint = name,
-                    "auto-detected prompt template"
-                );
-                return Self::ExaoneChat;
-            }
-            if lower.contains("minicpm") {
-                tracing::debug!(
-                    template = "MiniCPMChat",
-                    source = "tokenizer_name",
-                    hint = name,
-                    "auto-detected prompt template"
-                );
-                return Self::MiniCPMChat;
-            }
-            if lower.contains("smollm") || lower.contains("smol-lm") {
-                tracing::debug!(
-                    template = "SmolLMChat",
-                    source = "tokenizer_name",
-                    hint = name,
-                    "auto-detected prompt template"
-                );
-                return Self::SmolLMChat;
-            }
-            if lower.contains("instruct") {
-                tracing::debug!(
-                    template = "Instruct",
-                    source = "tokenizer_name",
-                    hint = name,
-                    "auto-detected prompt template"
-                );
-                return Self::Instruct;
-            }
-        }
-
-        // Priority 3: Fallback ΓÇö no recognisable signature found
-        tracing::warn!(template = "Raw", "no template signature found; falling back to Raw");
-        Self::Raw
+        detection::detect(tokenizer_name, chat_template_jinja)
     }
 
     /// Apply the template to a user prompt
